--- a/fe/fe-core/src/main/java/org/apache/doris/common/profile/SummaryProfile.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/common/profile/SummaryProfile.java
@@ -75,6 +75,7 @@ public class SummaryProfile {
     public static final String GET_PARTITION_VERSION_COUNT = "Get Partition Version Count";
     public static final String GET_PARTITION_VERSION_BY_HAS_DATA_COUNT = "Get Partition Version Count (hasData)";
 
+    public static final String PARSE_SQL_TIME = "Parse SQL Time";
     public static final String NEREIDS_ANALYSIS_TIME = "Nereids Analysis Time";
     public static final String NEREIDS_REWRITE_TIME = "Nereids Rewrite Time";
     public static final String NEREIDS_OPTIMIZE_TIME = "Nereids Optimize Time";
@@ -87,7 +88,7 @@ public class SummaryProfile {
             START_TIME, END_TIME, TOTAL_TIME, TASK_STATE, USER, DEFAULT_DB, SQL_STATEMENT);
 
     public static final ImmutableList<String> EXECUTION_SUMMARY_KEYS = ImmutableList.of(
-            NEREIDS_ANALYSIS_TIME, NEREIDS_REWRITE_TIME, NEREIDS_OPTIMIZE_TIME, NEREIDS_TRANSLATE_TIME,
+            PARSE_SQL_TIME, NEREIDS_ANALYSIS_TIME, NEREIDS_REWRITE_TIME, NEREIDS_OPTIMIZE_TIME, NEREIDS_TRANSLATE_TIME,
             WORKLOAD_GROUP, ANALYSIS_TIME,
             PLAN_TIME, JOIN_REORDER_TIME, CREATE_SINGLE_NODE_TIME, QUERY_DISTRIBUTED_TIME,
             INIT_SCAN_NODE_TIME, FINALIZE_SCAN_NODE_TIME, GET_SPLITS_TIME, GET_PARTITIONS_TIME,
@@ -120,6 +121,8 @@ public class SummaryProfile {
     private RuntimeProfile summaryProfile;
     private RuntimeProfile executionSummaryProfile;
 
+    private long parseSqlStartTime = -1;
+    private long parseSqlFinishTime = -1;
     private long nereidsAnalysisFinishTime = -1;
     private long nereidsRewriteFinishTime = -1;
     private long nereidsOptimizeFinishTime = -1;
@@ -204,6 +207,8 @@ public class SummaryProfile {
     }
 
     private void updateExecutionSummaryProfile() {
+        executionSummaryProfile.addInfoString(PARSE_SQL_TIME, getPrettyParseSqlTime());
+        executionSummaryProfile.addInfoString(NEREIDS_ANALYSIS_TIME, getPrettyNereidsAnalysisTime());
         executionSummaryProfile.addInfoString(NEREIDS_ANALYSIS_TIME, getPrettyNereidsAnalysisTime());
         executionSummaryProfile.addInfoString(NEREIDS_REWRITE_TIME, getPrettyNereidsRewriteTime());
         executionSummaryProfile.addInfoString(NEREIDS_OPTIMIZE_TIME, getPrettyNereidsOptimizeTime());
@@ -232,6 +237,14 @@ public class SummaryProfile {
             executionSummaryProfile.addInfoString(GET_PARTITION_VERSION_BY_HAS_DATA_COUNT,
                     getPrettyGetPartitionVersionByHasDataCount());
         }
+    }
+
+    public void setParseSqlStartTime(long parseSqlStartTime) {
+        this.parseSqlStartTime = parseSqlStartTime;
+    }
+
+    public void setParseSqlFinishTime(long parseSqlFinishTime) {
+        this.parseSqlFinishTime = parseSqlFinishTime;
     }
 
     public void setNereidsAnalysisTime() {
@@ -441,28 +454,35 @@ public class SummaryProfile {
         }
     }
 
-    private String getPrettyNereidsAnalysisTime() {
+    public String getPrettyParseSqlTime() {
+        if (parseSqlStartTime == -1 || parseSqlFinishTime == -1) {
+            return "N/A";
+        }
+        return RuntimeProfile.printCounter(parseSqlFinishTime - parseSqlStartTime, TUnit.TIME_MS);
+    }
+
+    public String getPrettyNereidsAnalysisTime() {
         if (nereidsAnalysisFinishTime == -1 || queryAnalysisFinishTime == -1) {
             return "N/A";
         }
         return RuntimeProfile.printCounter(nereidsAnalysisFinishTime - queryBeginTime, TUnit.TIME_MS);
     }
 
-    private String getPrettyNereidsRewriteTime() {
+    public String getPrettyNereidsRewriteTime() {
         if (nereidsRewriteFinishTime == -1 || nereidsAnalysisFinishTime == -1) {
             return "N/A";
         }
         return RuntimeProfile.printCounter(nereidsRewriteFinishTime - nereidsAnalysisFinishTime, TUnit.TIME_MS);
     }
 
-    private String getPrettyNereidsOptimizeTime() {
+    public String getPrettyNereidsOptimizeTime() {
         if (nereidsOptimizeFinishTime == -1 || nereidsRewriteFinishTime == -1) {
             return "N/A";
         }
         return RuntimeProfile.printCounter(nereidsOptimizeFinishTime - nereidsRewriteFinishTime, TUnit.TIME_MS);
     }
 
-    private String getPrettyNereidsTranslateTime() {
+    public String getPrettyNereidsTranslateTime() {
         if (nereidsTranslateFinishTime == -1 || nereidsOptimizeFinishTime == -1) {
             return "N/A";
         }