--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/CustomRewriteJob.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/CustomRewriteJob.java
@@ -17,6 +17,8 @@
 
 package org.apache.doris.nereids.jobs.rewrite;
 
+import org.apache.doris.nereids.CascadesContext;
+import org.apache.doris.nereids.PlanProcess;
 import org.apache.doris.nereids.jobs.Job;
 import org.apache.doris.nereids.jobs.JobContext;
 import org.apache.doris.nereids.rules.RuleType;
@@ -52,7 +54,8 @@ public class CustomRewriteJob implements RewriteJob {
         if (disableRules.contains(ruleType.type())) {
             return;
         }
-        Plan root = context.getCascadesContext().getRewritePlan();
+        CascadesContext cascadesContext = context.getCascadesContext();
+        Plan root = cascadesContext.getRewritePlan();
         // COUNTER_TRACER.log(CounterEvent.of(Memo.get=-StateId(), CounterType.JOB_EXECUTION, group, logicalExpression,
         //         root));
         Plan rewrittenRoot = customRewriter.get().rewriteRoot(root, context);
@@ -62,12 +65,14 @@ public class CustomRewriteJob implements RewriteJob {
 
         // don't remove this comment, it can help us to trace some bug when developing.
 
-        // if (!root.deepEquals(rewrittenRoot)) {
-        //     String traceBefore = root.treeString();
-        //     String traceAfter = root.treeString();
-        //     printTraceLog(ruleType, traceBefore, traceAfter);
-        // }
-        context.getCascadesContext().setRewritePlan(rewrittenRoot);
+        if (!root.deepEquals(rewrittenRoot)) {
+            if (cascadesContext.showPlanProcess()) {
+                PlanProcess planProcess = new PlanProcess(
+                        ruleType.name(), root.treeString(), rewrittenRoot.treeString());
+                cascadesContext.addPlanProcess(planProcess);
+            }
+        }
+        cascadesContext.setRewritePlan(rewrittenRoot);
     }
 
     @Override