--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/commands/ExplainCommand.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/commands/ExplainCommand.java
@@ -61,11 +61,13 @@ public class ExplainCommand extends Command implements NoForward {
 
     private final ExplainLevel level;
     private final LogicalPlan logicalPlan;
+    private final boolean showPlanProcess;
 
-    public ExplainCommand(ExplainLevel level, LogicalPlan logicalPlan) {
+    public ExplainCommand(ExplainLevel level, LogicalPlan logicalPlan, boolean showPlanProcess) {
         super(PlanType.EXPLAIN_COMMAND);
         this.level = level;
         this.logicalPlan = logicalPlan;
+        this.showPlanProcess = showPlanProcess;
     }
 
     @Override
@@ -76,7 +78,8 @@ public class ExplainCommand extends Command implements NoForward {
         }
         explainPlan = ((LogicalPlan) ((Explainable) logicalPlan).getExplainPlan(ctx));
         LogicalPlanAdapter logicalPlanAdapter = new LogicalPlanAdapter(explainPlan, ctx.getStatementContext());
-        logicalPlanAdapter.setIsExplain(new ExplainOptions(level));
+        ExplainOptions explainOptions = new ExplainOptions(level, showPlanProcess);
+        logicalPlanAdapter.setIsExplain(explainOptions);
         executor.setParsedStmt(logicalPlanAdapter);
         NereidsPlanner planner = new NereidsPlanner(ctx.getStatementContext());
         if (ctx.getSessionVariable().isEnableMaterializedViewRewrite()) {
@@ -85,7 +88,11 @@ public class ExplainCommand extends Command implements NoForward {
         planner.plan(logicalPlanAdapter, ctx.getSessionVariable().toThrift());
         executor.setPlanner(planner);
         executor.checkBlockRules();
-        executor.handleExplainStmt(planner.getExplainString(new ExplainOptions(level)), true);
+        if (showPlanProcess) {
+            executor.handleExplainPlanProcessStmt(planner.getCascadesContext().getPlanProcesses());
+        } else {
+            executor.handleExplainStmt(planner.getExplainString(explainOptions), true);
+        }
     }
 
     @Override
@@ -99,5 +106,9 @@ public class ExplainCommand extends Command implements NoForward {
 
     public LogicalPlan getLogicalPlan() {
         return logicalPlan;
+    }
+
+    public boolean showPlanProcess() {
+        return showPlanProcess;
     }
 }