… fe/fe-core/src/main/antlr4/org/apache/doris/nereids/DorisLexer.g4
… fe/fe-core/src/main/java/org/apache/doris/nereids/jobs/rewrite/RootPlanTreeRewriteJob.java
… fe/fe-core/src/main/java/org/apache/doris/qe/ConnectProcessor.java
… fe/fe-core/src/main/java/org/apache/doris/qe/StmtExecutor.java
… fe/fe-core/src/test/java/org/apache/doris/nereids/util/PlanChecker.java